BrainIAC/requirements.txt
pytorch-lightning==2.3.3
monai==1.3.2
nibabel==5.2.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
numpy==1.23.5
autograd==1.7.0
matplotlib==3.7.1
SimpleITK==2.4.0
tqdm
pydicom
wandb
lifelines
opencv-python